pyproject.toml
[project]
name = "flask-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-restful>=0.3.10",
    "pillow>=12.0.0",
    "pytest>=9.0.1",
    "requests>=2.32.5",
    "pyodbc>=5.0.0",
    "python-dotenv>=1.0.0",
]
